Use ast module instead of compiler module for parsing files.
This CL is part of the work getting GYP to run under Python 3.
The compiler module we used to parse GYP files was removed in
Python 3, so this CL switches to use the ast module instead,
which exists in both Python 2 and Python 3.

This change is not sufficient to run under Python 3; there's
more work remaining. Things should still work under Python 2
just fine.

This work is derived from Ashley.Whetter@gmail.com's
original work in https://codereview.chromium.org/1454433002/.

Bug: gyp:36

Change-Id: I9a9835560491c3d8cd5426623484dc4a46af1d86
Reviewed-on: https://chromium-review.googlesource.com/c/1360352
Reviewed-by: Mark Mentovai <mark@chromium.org>