Fixes needed for Python3 on Win32.
This CL contains the remaining fixes needed for the tests to
pass when running under Python 3.7 on Win10. Mostly these
are just changes to convert the output of subprocesses to
strings, but there are a few other minor compatibility fixes
as well.

With this CL, Python 3 should work everywhere we support it.
Python 2 continues to work as well.

Bug: gyp:36
Change-Id: Ie696a1845fb2dba4d86929be886becc5d300ef7a
Reviewed-on: https://chromium-review.googlesource.com/c/1365010
Reviewed-by: Mark Mentovai <mark@chromium.org>