Add SpectreMitigation attribute for msvs
This CL allows gyp to recognize the SpectreMitigation msvs_attribute
and add it to the generated vcxproj file for MSBuild.
Possible values for the attribute are Spectre, SpectreLoad,
SpectreLoadCF, and false.

The /Qspectre compiler option is not enough to add
full Spectre mitigation, because even though it causes MSBuild
to add additional instructions to the generated object files,
it does not cause MSBuild to link against Spectre-mitigated
libraries provided by Visual Studio.

Change-Id: I8294395659354f3ce60a355622da4b026abcceba
Reviewed-on: https://chromium-review.googlesource.com/c/external/gyp/+/4265144
Reviewed-by: Mark Mentovai <mark@chromium.org>